# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mentor IA API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserConfig {
  id: number;
  name: string;
  prioritySubjects: string[];
  createdAt: string;
}

export interface UpdateUserConfigBody {
  name: string;
  prioritySubjects?: string[];
}

export type SessionCategory =
  (typeof SessionCategory)[keyof typeof SessionCategory];

export const SessionCategory = {
  enem: "enem",
  programacao: "programacao",
} as const;

export interface Session {
  id: number;
  subject: string;
  category: SessionCategory;
  durationMinutes: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  quality: number;
  notes?: string | null;
  createdAt: string;
}

export type CreateSessionBodyCategory =
  (typeof CreateSessionBodyCategory)[keyof typeof CreateSessionBodyCategory];

export const CreateSessionBodyCategory = {
  enem: "enem",
  programacao: "programacao",
} as const;

export interface CreateSessionBody {
  subject: string;
  category: CreateSessionBodyCategory;
  /** @minimum 5 */
  durationMinutes: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  quality: number;
  notes?: string | null;
}

export type SessionStatsBySubjectItem = {
  subject: string;
  category: string;
  totalMinutes: number;
};

export type SessionStatsLast7DaysItem = {
  date: string;
  totalMinutes: number;
};

export interface SessionStats {
  totalMinutesToday: number;
  totalMinutesWeek: number;
  consistencyPercent: number;
  streak: number;
  bySubject: SessionStatsBySubjectItem[];
  last7Days: SessionStatsLast7DaysItem[];
}

export interface Habit {
  id: number;
  date: string;
  sleepHours?: number | null;
  exerciseDone: boolean;
  meditationDone: boolean;
  waterLiters?: number | null;
  /**
   * @minimum 1
   * @maximum 5
   */
  mood?: number | null;
  createdAt: string;
}

export interface CreateHabitBody {
  date: string;
  sleepHours?: number | null;
  exerciseDone?: boolean;
  meditationDone?: boolean;
  waterLiters?: number | null;
  /**
   * @minimum 1
   * @maximum 5
   */
  mood?: number | null;
}

export type QuestionResult =
  (typeof QuestionResult)[keyof typeof QuestionResult];

export const QuestionResult = {
  correct: "correct",
  wrong: "wrong",
  skipped: "skipped",
} as const;

export interface Question {
  id: number;
  subject: string;
  topic?: string | null;
  result: QuestionResult;
  /**
   * @minimum 1
   * @maximum 5
   */
  difficulty?: number | null;
  source?: string | null;
  createdAt: string;
}

export type CreateQuestionBodyResult =
  (typeof CreateQuestionBodyResult)[keyof typeof CreateQuestionBodyResult];

export const CreateQuestionBodyResult = {
  correct: "correct",
  wrong: "wrong",
  skipped: "skipped",
} as const;

export interface CreateQuestionBody {
  subject: string;
  topic?: string | null;
  result: CreateQuestionBodyResult;
  /**
   * @minimum 1
   * @maximum 5
   */
  difficulty?: number | null;
  source?: string | null;
}

export interface PaginatedQuestions {
  data: Question[];
  total: number;
  page: number;
  limit: number;
}

export interface QuestionSubjectStat {
  subject: string;
  total: number;
  correct: number;
  wrong: number;
  skipped: number;
  accuracyPercent: number;
}

export interface Goal {
  id: number;
  subject?: string | null;
  targetMinutes: number;
  weekStart: string;
  achieved: boolean;
  actualMinutes: number;
}

export interface CreateGoalBody {
  subject?: string | null;
  targetMinutes: number;
  weekStart: string;
}

export interface DashboardData {
  sessionStats: SessionStats;
  todayHabits?: Habit | null;
  questionStats: QuestionSubjectStat[];
  goals: Goal[];
  recentSessions: Session[];
  medicineIndex: number;
  codeIndex: number;
  streak: number;
  userName: string;
  latestInsight?: string | null;
}

export type MentorInsightType =
  (typeof MentorInsightType)[keyof typeof MentorInsightType];

export const MentorInsightType = {
  daily: "daily",
  decision: "decision",
  weekly_review: "weekly_review",
} as const;

export interface MentorInsight {
  insight: string;
  type: MentorInsightType;
}

export type MentorChatBodyMessagesItemRole =
  (typeof MentorChatBodyMessagesItemRole)[keyof typeof MentorChatBodyMessagesItemRole];

export const MentorChatBodyMessagesItemRole = {
  user: "user",
  assistant: "assistant",
} as const;

export type MentorChatBodyMessagesItem = {
  role: MentorChatBodyMessagesItemRole;
  content: string;
};

export interface MentorChatBody {
  messages: MentorChatBodyMessagesItem[];
}

export type ListSessionsParams = {
  days?: number;
  subject?: string;
  category?: ListSessionsCategory;
};

export type ListSessionsCategory =
  (typeof ListSessionsCategory)[keyof typeof ListSessionsCategory];

export const ListSessionsCategory = {
  enem: "enem",
  programacao: "programacao",
} as const;

export type GetSessionStatsParams = {
  days?: number;
};

export type ListHabitsParams = {
  days?: number;
};

export type ListQuestionsParams = {
  subject?: string;
  days?: number;
  result?: ListQuestionsResult;
  page?: number;
  limit?: number;
};

export type ListQuestionsResult =
  (typeof ListQuestionsResult)[keyof typeof ListQuestionsResult];

export const ListQuestionsResult = {
  correct: "correct",
  wrong: "wrong",
  skipped: "skipped",
} as const;