# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mentor IA API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /habits:
    get:
      operationId: listHabits
      tags: [habits]
      summary: List daily habits
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: List of habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      operationId: createOrUpdateHabit
      tags: [habits]
      summary: Create or update today's habit record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHabitBody"
      responses:
        "200":
          description: Upserted habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
  /questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions
      parameters:
        - name: subject
          in: query
          schema:
            type: string
        - name: days
          in: query
          schema:
            type: integer
            default: 30
        - name: result
          in: query
          schema:
            type: string
            enum: [correct, wrong, skipped]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedQuestions"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Record a question answer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Created question record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
...
components:
  schemas:
...
    Habit:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        sleepHours:
          type: number
          nullable: true
        exerciseDone:
          type: boolean
        meditationDone:
          type: boolean
        waterLiters:
          type: number
          nullable: true
        mood:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        createdAt:
          type: string
      required: [id, date, exerciseDone, meditationDone, createdAt]
    CreateHabitBody:
      type: object
      properties:
        date:
          type: string
        sleepHours:
          type: number
          nullable: true
        exerciseDone:
          type: boolean
        meditationDone:
          type: boolean
        waterLiters:
          type: number
          nullable: true
        mood:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
      required: [date]
    Question:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        topic:
          type: string
          nullable: true
        result:
          type: string
          enum: [correct, wrong, skipped]
        difficulty:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        source:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, subject, result, createdAt]
    CreateQuestionBody:
      type: object
      properties:
        subject:
          type: string
        topic:
          type: string
          nullable: true
        result:
          type: string
          enum: [correct, wrong, skipped]
        difficulty:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        source:
          type: string
          nullable: true
      required: [subject, result]
    PaginatedQuestions:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
...