# lib/api-zod/src/generated/types/createHabitBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mentor IA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateHabitBody {
  date: string;
  sleepHours?: number | null;
  exerciseDone?: boolean;
  meditationDone?: boolean;
  waterLiters?: number | null;
  /**
   * @minimum 1
   * @maximum 5
   */
  mood?: number | null;
}