# lib/api-zod/src/generated/types/createQuestionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mentor IA API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateQuestionBodyResult } from "./createQuestionBodyResult";

export interface CreateQuestionBody {
  subject: string;
  topic?: string | null;
  result: CreateQuestionBodyResult;
  /**
   * @minimum 1
   * @maximum 5
   */
  difficulty?: number | null;
  source?: string | null;
}