# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mentor IA API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionResult } from "./questionResult";

export interface Question {
  id: number;
  subject: string;
  topic?: string | null;
  result: QuestionResult;
  /**
   * @minimum 1
   * @maximum 5
   */
  difficulty?: number | null;
  source?: string | null;
  createdAt: string;
}